+++ CHANGELOG.md
@@ -7,6 +7,7 @@ Changes:
 - Allow types to inherit from multiple base classes/interfaces
 - Cleanup SchemaType to use an enum instead of lots of boolean type variables. Previous constructor is obsolete
 - Cleanup Interfaces api - added a `AddAllBaseTypes`, `AddBaseType` and `AddBaseType(string)` which provides a lot more flexiblity. See updated docs
+- Added support for Inline Fragments for types that have interfaces
 
 Fixes: